practix3/images/movies-admin/requirements.txt
asgiref==3.8.1
    # via django
attrs==23.2.0
    # via
    #   jsonschema
    #   referencing
django==5.0.3
    # via
    #   -r requirements.in
    #   django-debug-toolbar
    #   django-filter
    #   djangorestframework
    #   drf-spectacular
django-debug-toolbar==4.3.0
    # via -r requirements.in
django-filter==24.2
    # via -r requirements.in
djangorestframework==3.15.1
    # via
    #   -r requirements.in
    #   drf-spectacular
drf-spectacular==0.27.1
    # via -r requirements.in
flake8==7.0.0
    # via
    #   -r requirements.in
    #   pep8-naming
inflection==0.5.1
    # via drf-spectacular
isort==5.13.2
    # via -r requirements.in
jsonschema==4.21.1
    # via drf-spectacular
jsonschema-specifications==2023.12.1
    # via jsonschema
mccabe==0.7.0
    # via flake8
pep8-naming==0.13.3
    # via -r requirements.in
psycopg2-binary==2.9.9
    # via -r requirements.in
pycodestyle==2.11.1
    # via flake8
pyflakes==3.2.0
    # via flake8
pyyaml==6.0.1
    # via drf-spectacular
referencing==0.34.0
    # via
    #   jsonschema
    #   jsonschema-specifications
rpds-py==0.18.0
    # via
    #   jsonschema
    #   referencing
sqlparse==0.4.4
    # via
    #   django
    #   django-debug-toolbar
typing-extensions==4.10.0
    # via asgiref
uritemplate==4.1.1
    # via drf-spectacular
